[DOCS] Adds a philosophy section to the guide
One of the main things that makes a good modifier is if it is able
to completely close over a concept or piece of functionality, and
effectively turn a side effect into derived state. I think that's a
bit of a high level concept to teach, but it's also really useful.
I tried to capture that in this section, without going into _too_ much
detail with what we mean by "side effect" and "derived state" and all
these fancy terms and such. I also tried to make it clear that it's ok
to write modifiers that _don't_ do this, just that it's a good thing
to aim for, especially in general purpose modifiers.

In the future, this could be part of the Ember guides in a more
advanced section possibly, after the more introductory stuff about
how to write them and which basic ones are included OOTB.