Update dependencies to fix issues with `@ember/destroyable`.
Since the original refactor to use `@ember/destroyable` landed, a new
version of `ember-cli-babel` was released that was **not** compatible
with our usage of `@ember/destroyable`. A new patch release was done to
fix that (in ember-cli-babel@7.22.1) but it is still possible for
consumers of ours to end up with the older version and have issues.

This bumps the minimum version of ember-cli-babel that we use to 7.22.1
and also updates the ember-destroyable-polyfill to at least 1.0.2 in
order to fix these isssues.