config: update TS and TS test matrix
- Use *latest* TypeScript 4.0 for developing the addon.
- Update the ember-try config to include 4.0 explicitly in the test
  matrix, while maintaining support for 3.6, 3.7, 3.9, and `next`.
- Use `ember-cli-typescript` 4.0 release: combined with latest
  `ember-cli-babel`, it makes everything work nicely on TS 3.9 and
  4.0. We're not currently using the features where it matters, but it
  keeps things tidy.
- Create a shared `tsconfig.json` to be used by both the *current* and
  *next* version of TypeScript, which does *not* use down-leveled types.

Previously, we were depending on everything to keep type-checking just
by dint of the normal CI pass, but weren't explicitly checking the type
tests against 3.9. Whoops. 😬 This was "fine" (as in this-is-fine.gif)
because we were checking against TS@next, but we need to make sure we
run the type tests against the *current* TS version as well as earlier
versions and `next`.