Fix issue with Ember 3.20.4+.
Prior to this change, we happened to be relying on accidental behavior
in Ember. Specifically, that the first argument of `createModifier`
would contain the `owner` property. This was not part of the RFC and is
not a documented feature. The correct way to obtain the owner has always
been by way of the `setModifierManager` API which this commit changes
to.