CI: fix package caching and simplify Node setup
The previous setup for this was reusing the package manager cache
across unrelated CI runs, which resulted in the Floating Dependencies
job "clobbering" all runs which happened to come after it. Drop all the
extra config in favor of a setup which more closely matches [the new
defaults for addons][1], and switch to the non-beta release of the
`setup-node` GitHub Action as well.

[1]: https://github.com/ember-cli/ember-addon-output/blob/41cd995bce3b9ad0f41885a43f473a9545dff9b1/.github/workflows/ci.yml#L21