Remove babel.config.js
Related to #101 and #102. Bumping to latest Embroider internals exposed
some further issue with Babel grabbing the `babel.config.js` used to
configure ESLint via `@babel/eslint-parser`. This is a temporary
workaround to unblock CI to work more generally, and which is "fine"
for now because we don't actually have any unusual syntax in this repo.
We will need to do *some* work upstream to finish fixing that more
generally for the ecosystem as a whole, though.