Deprecate `this.element` and `this.args` on class-based modifiers
The introduction of the `modify` hook on class-based modifiers allows
users to use the `element` and both `positional` and `named` arguments
directly, as arguments to `modify`, rather than having them as the
class properties `this.element` and `this.args` (see #217).
Accordingly, deprecate access to both `this.element` and `this.args`.

(Users who want to stash the `element` on their modifier class can
continue to do so directly, by assigning it from the `modify` call.)