Fix "default" positional and named argument types
Previously, we resolved the default types of positional and named
arguments to `unknown` and `Record<string, unknown>` respectively in TS
context. However, Glint needs those to be `[]` (an empty tuple) and a
special `EmptyObject` type respectively, so that it can show errors
when users pass positional parameters to a type where.

This *only* applies to new `Signature` types. It does not change the
the previous defaults of `unknown[]` and `Record<string, unknown>` when
leaving the types unspecified entirely, and does not affect inference
when using non-`Signature` type params explicitly.

Note: this default *will* be changed in v4.0 so *all* unspecified named
and positional arguments get the new defaults, so they work correctly
with Glint.