Use "export type" for FunctionBasedModifier interface
Without this change you get a broken reexport when built by babel, because it doesn't know to strip FunctionBasedModifier out of index.ts, but it does know to strip it out of -private/function-based/modifier.ts.

This results in a warning from webpack when using embroider. It's also technically a SyntaxError to try to reexport a thing that doesn't exist.